+++ src/main/java/io/github/astrapi69/crypt/data/certificate/ExtensionInfo.java
@@ -35,6 +35,7 @@ import lombok.experimental.FieldDefaults;
  * @deprecated use instead the class {@link io.github.astrapi69.crypt.data.model.ExtensionInfo}.
  *             Note will be removed in next minor version
  */
+@Deprecated
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class ExtensionInfo
 {