+++ src/main/java/io/github/astrapi69/crypt/data/model/PemKeyModel.java
@@ -38,6 +38,7 @@ import lombok.experimental.SuperBuilder;
  * @deprecated use instead the class {@link io.github.astrapi69.crypt.data.model.PemKeyInfo}. Note
  *             will be removed in next minor version
  */
+@Deprecated
 @Data
 @RequiredArgsConstructor
 @SuperBuilder(toBuilder = true)