+++ src/test/java/io/github/astrapi69/crypt/data/key/reader/PrivateKeyReaderTest.java
@@ -161,7 +161,6 @@ public class PrivateKeyReaderTest
 		assertEquals(algorithm, "RSASSA-PSS");
 	}
 
-
 	/**
 	 * Test method for {@link PrivateKeyReader#getKeyFormat(File)}
 	 */