+++ src/test/java/io/github/astrapi69/crypt/data/model/DistinguishedNameInfoTest.java
@@ -24,7 +24,7 @@
  */
 package io.github.astrapi69.crypt.data.model;
 
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import org.junit.jupiter.api.Test;